Move verify_all_files from raw filestore into CAS
